openclaw-trader/requirements.txt
anthropic>=0.40.0
pandas>=2.0.0
numpy>=1.24.0
requests>=2.31.0
websocket-client>=1.6.0
python-telegram-bot>=21.0
python-dotenv>=1.0.0
apscheduler>=3.10.0
pydantic>=2.0.0
